Rename all 'util' modules to 'utils'.
This matches the convention used throughout the rest of the codebase. Note that
this is potentially a breaking change for some users but one which is probably
worth making in this pre-release state.
