+++ rig/place_and_route/allocate/greedy.py
@@ -21,7 +21,7 @@ from ..constraints import ReserveResourceConstraint, AlignResourceConstraint
 
 from ..exceptions import InsufficientResourceError
 
-from .util import slices_overlap, align
+from .utils import slices_overlap, align
 
 
 def allocate(vertices_resources, nets, machine, constraints, placements):
