+++ rig/place_and_route/route/tests/test_generic.py
@@ -14,7 +14,7 @@ from rig.place_and_route.constraints import RouteEndpointConstraint
 
 from rig.place_and_route.exceptions import MachineHasDisconnectedSubregion
 
-from rig.place_and_route.route.util import links_between
+from rig.place_and_route.route.utils import links_between
 
 from rig.place_and_route import route as default_route
 from rig.place_and_route.route.ner import route as ner_route
