+++ rig/place_and_route/tests/test_utils.py
@@ -1,4 +1,4 @@
-from rig.place_and_route.util import build_application_map, \
+from rig.place_and_route.utils import build_application_map, \
     build_routing_tables
 
 from rig.machine import Cores
